excel_analyze/requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
openpyxl>=3.1.2
pyvis>=0.3.2
pyyaml>=6.0
